requirements.txt
discord.py
aiohttp
python-dotenv
pyngrok
libsql-client
libsql-client
